+++ fe/src/pages/adminPages/Updating/UpdateLecturer.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react"
+import { useEffect, useMemo, useState } from "react"
 import { useNavigate, useParams } from "react-router-dom"
 import { base } from "../../../App"
 
@@ -24,7 +24,7 @@ export default function UpdateLecturer() {
   const navigate = useNavigate()
   const fetchLecturer = async () => {
     try {
-      let url = base+`/lecturers?id=${id}`
+      let url = base + `/lecturers?id=${id}`
       const res = await fetch(url);
       const data = await res.json()
       if (data.lecturer.length) {
@@ -32,8 +32,8 @@ export default function UpdateLecturer() {
       } else {
         throw new Error('Lecturer not found')
       }
-    } catch (err:any) {
-      alert(err?.message||'something went wrong')
+    } catch (err: any) {
+      alert(err?.message || 'something went wrong')
       navigate(-1)
     }
   }
@@ -46,32 +46,35 @@ export default function UpdateLecturer() {
   }
   const updateLecturer = async () => {
     try {
-      const f = new FormData()
-      f.append('firstName', form.firstName)
-      f.append('lastName', form.lastName)
-      f.append('otherNames', form.otherNames)
-      f.append('email', form.email)
-      f.append('phone', form.phone)
-      f.append('dob', form.dob)
-      f.append('gender', form.gender)
-      f.append('password', form.password)
-      f.append('id', form.id)
+      
 
-      let url = base+`/lecturers?id=${id}`
+      let url = base + `/lecturers?id=${id}`
       const res = await fetch(url, {
         method: 'PUT',
-        body: f
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify(form),
       })
+
       const data = await res.json()
-      console.log(data)
-    } catch (err:any) {
-      alert(err?.message||'something went wrong')
+      if(data.ok){
+        alert('Lecturer updated successfully')
+      }else{
+        throw new Error(data?.message || 'Error updating lecturer')
+      }
+    } catch (err: any) {
+      alert(err?.message || 'something went wrong')
     }
   }
   const delete_lecturer = () => {
     const reply = prompt('are you sure you want to delete this Lecturer? Type yes to confirm')
     if (reply?.toLowerCase() !== 'yes') return
-    fetch(base+'/lecturers?id=' + id, {
+    if(!id){ 
+      alert('no lecturer id found')
+      return
+    }
+    fetch(base + '/lecturers?id=' + id, {
       method: 'DELETE',
       headers: {
         'Content-Type': 'application/json'
@@ -92,6 +95,19 @@ export default function UpdateLecturer() {
       })
   }
 
+  const dob = useMemo(() => {
+    if (form.dob) {
+     /* format to yyyy-mm-dd */
+      const date = new Date(form.dob)
+      const year = date.getFullYear()
+      const month = date.getMonth() + 1
+      const day = date.getDate()
+      return `${year}-${month < 10 ? '0' + month : month}-${day < 10 ? '0' + day : day}`
+    }else{
+      return ''
+    }
+  }, [form.dob])
+
   return (
     <div className="p-4 flex flex-col w-full items-center h-[90vh] overflow-y-auto">
       <h1>{form.id}</h1>
@@ -99,7 +115,7 @@ export default function UpdateLecturer() {
         <label htmlFor="email">Email</label>
         <input type="text" name="email" id="label" value={form.email} onChange={(e) => setForm({ ...form, email: e.target.value })} className="text-[#346237] h-[40px] bg-transparent border border-[#346837] rounded-[5px] px-2" />
         <label htmlFor="fN">First Name</label>
-        <input type="text" name="firstName"  id='fN' value={form.firstName} onChange={(e) => setForm({ ...form, firstName: e.target.value })} className="text-[#346237] h-[40px] bg-transparent border border-[#346837] rounded-[5px] px-2" />
+        <input type="text" name="firstName" id='fN' value={form.firstName} onChange={(e) => setForm({ ...form, firstName: e.target.value })} className="text-[#346237] h-[40px] bg-transparent border border-[#346837] rounded-[5px] px-2" />
         <label htmlFor="lN">Last Name</label>
         <input type="text" id="lN" name="name" value={form.lastName} onChange={(e) => setForm({ ...form, lastName: e.target.value })} className="text-[#346237] h-[40px] bg-transparent border border-[#346837] rounded-[5px] px-2" />
         <label htmlFor="oN">Other Names</label>
@@ -111,7 +127,7 @@ export default function UpdateLecturer() {
         <label htmlFor="dA">Degree Acquired</label>
         <input type="text" id="dA" name="degree" value={form.degreeAcquired} onChange={(e) => setForm({ ...form, degreeAcquired: e.target.value })} className="text-[#346237] h-[40px] bg-transparent border border-[#346837] rounded-[5px] px-2" />
         <label htmlFor="dob">Date of Birth</label>
-        <input type="text" id="dob" name="dob" value={form.dob} onChange={(e) => setForm({ ...form, dob: e.target.value })} className="text-[#346237] h-[40px] bg-transparent border border-[#346837] rounded-[5px] px-2" />
+        <input type="date" id="dob" name="dob" value={dob} onChange={(e) => setForm({ ...form, dob: e.target.value })} className="text-[#346237] h-[40px] bg-transparent border border-[#346837] rounded-[5px] px-2" />
         <label htmlFor="pwd">Password</label>
         <input type="text" id="Pwd" name="password" value={form.password} onChange={(e) => setForm({ ...form, password: e.target.value })} className="text-[#346237] h-[40px] bg-transparent border border-[#346837] rounded-[5px] px-2" />
         <button className="bg-[#346837] py-2 mt-4 rounded-[4px] text-[#fff]">
